重构 AI 服务模块
